Rename observers to insight observers
