+++ app/Observers/Site/RegionInsightObserver.php
@@ -1,31 +1,25 @@
 <?php
 
-declare(strict_types=1);
-
 namespace App\Observers\Site;
 
 use App\Helpers\InsightsHelper;
-use App\Models\Site\Province;
+use App\Models\Site\Region;
 use App\Traits\AsInsightSender;
 use Illuminate\Database\Eloquent\Model;
 
-class ProvinceObserver
+class RegionInsightObserver
 {
     use AsInsightSender;
 
     /**
-     * @param  Province  $model
+     * @param  Region  $model
      */
     private function sendInsights(Model $model, bool $shouldIncrement): void
     {
-        $labels = [
-            'region' => $model->region->slug,
-        ];
-
         if ($shouldIncrement) {
-            InsightsHelper::incrementGauge('province_total', $labels);
+            InsightsHelper::incrementGauge('region_total');
         } else {
-            InsightsHelper::decrementGauge('province_total', $labels);
+            InsightsHelper::decrementGauge('region_total');
         }
     }
 }
