--- app/Observers/Site/RegionObserver.php
@@ -1,25 +1,0 @@
-<?php
-
-namespace App\Observers\Site;
-
-use App\Helpers\InsightsHelper;
-use App\Models\Site\Region;
-use App\Traits\AsInsightSender;
-use Illuminate\Database\Eloquent\Model;
-
-class RegionObserver
-{
-    use AsInsightSender;
-
-    /**
-     * @param  Region  $model
-     */
-    private function sendInsights(Model $model, bool $shouldIncrement): void
-    {
-        if ($shouldIncrement) {
-            InsightsHelper::incrementGauge('region_total');
-        } else {
-            InsightsHelper::decrementGauge('region_total');
-        }
-    }
-}
